Refactor of damage reports system. Also fixes an issue where blank damage report steps still appear in the report
