+++ server/classes/engine.js
@@ -1,4 +1,5 @@
-import { System, HeatMixin } from "./generic";
+import { System } from "./generic";
+import HeatMixin from "./generic/heatMixin";
 
 export default class Engine extends HeatMixin(System) {
   constructor(params = {}) {
