+++ server/classes/generic/damageReports/functions/damageTeam.js
