+++ server/classes/generic/damageReports/functions/exocomps.js
