+++ server/classes/generic/damageReports/functions/generic.js
