+++ server/classes/generic/damageReports/functions/longRangeMessage.js
