+++ server/classes/generic/damageReports/functions/power.js
