+++ server/classes/generic/damageReports/functions/probeLaunch.js
