+++ server/classes/phasers.js
@@ -1,4 +1,6 @@
-import { System, HeatMixin } from "./generic";
+import { System } from "./generic";
+import HeatMixin from "./generic/heatMixin";
+
 // It's easier to manage power if there is just one phaser system with multiple beams than
 // If there are multiple phaser systems
 export default class Phasers extends HeatMixin(System) {
