+++ server/classes/reactor.js
@@ -1,4 +1,5 @@
-import { System, HeatMixin } from "./generic";
+import { System } from "./generic";
+import HeatMixin from "./generic/heatMixin";
 
 export default class Reactor extends HeatMixin(System) {
   constructor(params) {
