+++ server/schema/types/system.js
@@ -38,91 +38,5 @@ type System implements SystemInterface {
 # Generic system type. Query any system by type.
 union SystemUnion = LRCommunications | Shield | Thruster | Engine | Transporter | Sensors | InternalComm
 
-type DamageStep {
-  id: ID
-  name: String
-  args: DamageStepArgs
-}
-
-type DamageStepArgs {
-  end: Boolean
-
-  #Damage Team Args
-  cleanup: Boolean
-  name: String
-  orders: String
-  room: String
-  preamble: String
-  type: String
-
-  #Damage Team Message Args
-  message: String
-
-  #Remote Access Args
-  code: String
-  backup: String
-
-  #Inventory Args
-  inventory: String
-
-  #Long Range Message Args
-  destination: String
-
-  #Probe Launch Args
-  equipment: String
-  query: String
-
-  #Generic Args
-
-  #Finish Args
-  reactivate: Boolean
-}
-
-input DamageStepInput {
-  id: ID
-  name: String
-  args: DamageStepArgsInput
-  type: DAMAGE_STEP_TYPES
-}
-
-enum DAMAGE_STEP_TYPES {
-  required
-  optional
-}
-
-input DamageStepArgsInput {
-  end: Boolean
-
-  #Damage Team Args
-  cleanup: Boolean
-  name: String
-  orders: String
-  room: String
-  preamble: String
-  type: String
-
-  #Damage Team Message Args
-  message: String
-
-  #Remote Access Args
-  code: String
-  backup: String
-  
-  #Inventory Args
-  inventory: String
-
-  #Long Range Message Args
-  destination: String
-
-  #Probe Launch Args
-  equipment: String
-  query: String
-
-  #Generic Args
-
-  #Finish Args
-  reactivate: Boolean
-}
-
 
 `;
